+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 Versions and bullets are arranged chronologically from latest to oldest.
 
+## Unreleased
+
+- Fix oddly parsed summary when an article is created.
+- Move from deprecated AddNewAccount hook to new LocalUserCreated hook. (https://github.com/femiwiki/DiscordNotifications/issues/2)
+- Localisation updates from https://translatewiki.net.
+
 ## 1.0.0-femiwiki
 
 ### Breaking changes
@@ -12,7 +18,7 @@ Versions and bullets are arranged chronologically from latest to oldest.
 ### Others
 
 - Support regular expression for denylisting pages.
-- Treat passed user input as plaintext. (https://github.com/kulttuuri/DiscordNotifications/issues/73)
+- ~~Treat passed user input as plaintext. (https://github.com/kulttuuri/DiscordNotifications/issues/73)~~
 - Format with Prettier.
 - Set up Codecov.